Extracted the FS filter dialog setup code to its own file.
